+++ src/cryptography/x509/__init__.py
@@ -40,6 +40,10 @@ from cryptography.x509.oid import (
     CertificatePoliciesOID, ExtendedKeyUsageOID, ExtensionOID, NameOID,
     ObjectIdentifier, SignatureAlgorithmOID, _SIG_OIDS_TO_HASH
 )
+from cryptography.x509.verification import (
+    CertificateVerificationContext, InvalidCertificate,
+    InvalidSigningCertificate
+)
 
 
 OID_AUTHORITY_INFORMATION_ACCESS = ExtensionOID.AUTHORITY_INFORMATION_ACCESS
@@ -131,6 +135,8 @@ __all__ = [
     "UnsupportedExtension",
     "ExtensionNotFound",
     "UnsupportedGeneralNameType",
+    "InvalidCertificate",
+    "InvalidSigningCertificate",
     "NameAttribute",
     "Name",
     "RelativeDistinguishedName",
@@ -175,6 +181,7 @@ __all__ = [
     "RevokedCertificateBuilder",
     "CertificateSigningRequestBuilder",
     "CertificateBuilder",
+    "CertificateVerificationContext",
     "Version",
     "_SIG_OIDS_TO_HASH",
     "OID_CA_ISSUERS",